+++ mlf_core/create/templates/mlflow/mlflow_pytorch/{{ cookiecutter.project_slug }}/{{ cookiecutter.project_slug_no_hyphen }}/{{ cookiecutter.project_slug_no_hyphen }}.py
@@ -71,7 +71,7 @@ def start_training(cuda, epochs, general_seed, pytorch_seed, log_interval,
         mlflow.pytorch.log_model(model, 'models')
 
         # Log hardware and software
-        log_sys_intel_conda_env('{{ cookiecutter.project_slug_no_hyphen }}')
+        log_sys_intel_conda_env()
 
         # Upload the TensorBoard event logs as a run artifact
         click.echo(click.style('Uploading TensorBoard events as a run artifact...', fg='blue'))