+++ mlf_core/lint/domains/mlflow.py
@@ -82,8 +82,7 @@ class MlflowPytorchLint(TemplateLinter, metaclass=GetLintingFunctionsMeta):
                                                   'torch.backends.cudnn.deterministic = True',
                                                   'torch.backends.cudnn.benchmark = False',
                                                   'set_general_random_seeds(general_seed)',
-                                                  'set_pytorch_random_seeds(pytorch_seed, use_cuda=use_cuda)',
-                                                  f'log_sys_intel_conda_env(\'{self.project_slug}\')']
+                                                  'set_pytorch_random_seeds(pytorch_seed, use_cuda=use_cuda)']
 
         for expected_line in expected_lines_pytorch_reproducibility:
             if expected_line not in project_slug_entry_point_content:
@@ -206,8 +205,7 @@ class MlflowTensorflowLint(TemplateLinter, metaclass=GetLintingFunctionsMeta):
                                                   'tf.config.threading.set_inter_op_parallelism_threads = 1  # CPU only',
                                                   'os.environ[\'TF_DETERMINISTIC_OPS\'] = \'1\'',
                                                   'set_general_random_seeds(general_seed)',
-                                                  'set_tensorflow_random_seeds(tensorflow_seed)',
-                                                  f'log_sys_intel_conda_env(\'{self.project_slug_no_hyphen}\')']
+                                                  'set_tensorflow_random_seeds(tensorflow_seed)']
 
         for expected_line in expected_lines_pytorch_reproducibility:
             if expected_line not in project_slug_entry_point_content:
@@ -289,8 +287,7 @@ class MlflowXGBoostLint(TemplateLinter, metaclass=GetLintingFunctionsMeta):
         expected_lines_xgboost_reproducibility = ['def set_xgboost_random_seeds(seed, param):',
                                                   'param[\'seed\'] = seed',
                                                   'set_general_random_seeds(general_seed)',
-                                                  'set_xgboost_random_seeds(xgboost_seed, param)',
-                                                  f'log_sys_intel_conda_env(\'{self.project_slug_no_hyphen}\')']
+                                                  'set_xgboost_random_seeds(xgboost_seed, param)']
 
         for expected_line in expected_lines_xgboost_reproducibility:
             if expected_line not in project_slug_entry_point_content:
@@ -394,8 +391,7 @@ class MlflowXGBoostDaskLint(TemplateLinter, metaclass=GetLintingFunctionsMeta):
         expected_lines_xgboost_reproducibility = ['def set_xgboost_dask_random_seeds(seed, param):',
                                                   'param[\'seed\'] = seed',
                                                   'set_general_random_seeds(general_seed)',
-                                                  'set_xgboost_dask_random_seeds(xgboost_seed, param)',
-                                                  f'log_sys_intel_conda_env(\'{self.project_slug_no_hyphen}\')']
+                                                  'set_xgboost_dask_random_seeds(xgboost_seed, param)']
 
         for expected_line in expected_lines_xgboost_reproducibility:
             if expected_line not in project_slug_entry_point_content: